+++ Adj/Algebra/Initial.hs
@@ -1,3 +1,0 @@
-module Adj.Algebra.Initial where
-
-data Initial