+++ Adj/Algebra/Morphism.hs
@@ -1,5 +1,0 @@
-module Adj.Algebra.Morphism (module Exports) where
-
-import Adj.Algebra.Morphism.Flat as Exports
-import Adj.Algebra.Morphism.Dual as Exports
-import Adj.Algebra.Morphism.Kleisli as Exports