+++ Adj/Algebra/Terminal.hs
@@ -1,3 +1,0 @@
-module Adj.Algebra.Terminal where
-
-data Terminal = Terminal