+++ Adj/Algebra/Unit.hs
@@ -1,3 +1,0 @@
-module Adj.Algebra.Unit where
-
-type family Unit (p :: * -> * -> *) = r | r -> p